// packages/crusher-electron-app/src/sdk/types.ts
export type KeyboardModifier = "Alt" | "Control" | "Meta" | "Shift";